modify newtodo page
